requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Werkzeug>=3.0.0
numpy>=1.26.0